style: mejor estilo para ver el producto, refactor: separar componentes como slideshow y ToolInfoModal, etc...
