+++ src/pages/toolPages/ToolMenu.css
@@ -83,3 +83,39 @@
 	background-color: transparent;
 	color: var(--text-primary);
 }
+
+.slideshow-container {
+	position: relative;
+	width: 200px;
+	height: 200px;
+	overflow: hidden;
+}
+.grupo {
+	display: flex;
+	justify-content: space-around;
+}
+
+.slideshow-image {
+	width: 100%;
+	height: 100%;
+}
+
+.nav-button {
+	position: absolute;
+	top: 50%;
+	transform: translateY(-50%);
+	z-index: 100;
+	cursor: pointer;
+	border: none;
+	background-color: rgba(0, 0, 0, 0.5);
+	padding: 10px;
+	color: white;
+}
+
+.nav-button.prev {
+	left: 0;
+}
+
+.nav-button.next {
+	right: 0;
+}
